# src/app/api/openapi.json/openapi-spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Preca API",
    "description": "API para la gestión de servicios de precalificación y screening en México",
    "version": "1.0.0",
    "contact": {
      "name": "Preca Support",
      "email": "soporte@preca.mx"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Servidor de desarrollo"
    }
  ],
...
  "paths": {
...
    "/webhooks/stripe": {
      "post": {
        "tags": [
          "Webhooks"
        ],
        "summary": "Stripe Webhook",
        "description": "Endpoint para recibir eventos de Stripe (checkout.session.completed, etc.). Verifica la firma de Stripe y procesa pagos.",
        "responses": {
          "200": {
            "description": "Evento procesado correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "received": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error de validación (firma inválida, payload malformado)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key",
        "description": "API Key de autenticación en formato: preca_xxxx_xxxxxxxxxxxxxxxx"
      },
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Cookie de sesión para usuarios autenticados vía web"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Mensaje de error"
          }
        },
        "required": [
          "error"
        ]
      },
...
    }
  }
}